feat: handle error using customError

+++ app/api/delete-review/[reviewID]/route.js
@@ -1,31 +1,37 @@
 import { NextResponse } from "next/server";
 import { connect } from "@/dbconfig/dbconfig";
 import Reviews from "@/models/reviewModel";
+import CustomError from "@/utils/errors";
 
-connect();
-//this [id ] router
 export async function DELETE(req, { params }) {
+  await connect();
   try {
     const reviewID = params.reviewID;
-    console.log(reviewID);
 
-    const deletedReview = await Reviews.findByIdAndDelete(reviewID);
+    if (!reviewID) {
+      throw new CustomError("review not found", 401);
+    }
 
-    console.log("going to revalidate");
-    if (deletedReview) {
-      console.log(deletedReview);
-      return NextResponse.json({
+    await Reviews.findByIdAndDelete(reviewID);
+
+    return NextResponse.json(
+      {
         message: "Review deleted successfully",
         success: true,
-      });
-    }
+      },
+      { status: 200 }
+    );
   } catch (error) {
-    console.log("error in review deleting route");
-    console.log(error);
-    return NextResponse.json({
-      error: error,
-      message: "Could not delete review",
-      success: false,
-    });
+    const statusCode = error.statusCode || 500;
+    const message = error.customMessage || "internal error";
+    console.log("error in delete-review\n", error);
+
+    return NextResponse.json(
+      {
+        message: message,
+        success: false,
+      },
+      { status: statusCode }
+    );
   }
 }
